backend/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Reduct backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.13.0",
    "gunicorn>=23.0.0",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "twelvelabs>=0.4.7",
    "uvicorn>=0.34.2",
]
